+++ src/api/apiHelper.ts
@@ -1,4 +1,30 @@
 import { getAppSetting } from '../getAppSetting';
+import axios from 'axios';
+
+const API_HOST = getAppSetting('REACT_APP_API_URL');
+
+const IN_DEV = process.env.NODE_ENV === 'development';
+
+export const climateApi = axios.create({
+  baseURL: API_HOST,
+});
+
+// Intercept and log out request config in dev
+climateApi.interceptors.request.use((config) => {
+  if (IN_DEV) {
+    console.log(config);
+  }
+  return config;
+});
+
+// Intercept and log out responses in dev
+climateApi.interceptors.response.use((config) => {
+  if (IN_DEV) {
+    console.log(config);
+  }
+  return config;
+});
+
 // Takes an api endpoint and optional query pramam then retruns the correct url to call on Climate Mind API
 export const buildUrl = (endpoint: string) => {
   const API_HOST = getAppSetting('REACT_APP_API_URL');
