+++ src/api/getQuestions.ts
@@ -1,14 +1,10 @@
-import axios from 'axios';
-import { buildUrl } from './apiHelper';
 import { TQuestions } from '../types/types';
+import { climateApi } from './apiHelper';
 
 export const getQuestions = async (): Promise<TQuestions> => {
-  const REQUEST_URL = buildUrl('/questions');
-
   try {
     // Call the api
-    console.log('REQEST', REQUEST_URL);
-    const response = await axios.get(REQUEST_URL);
+    const response = await climateApi.get('/questions');
     const data = response.data;
     return data;
   } catch (err) {
